tidy DropdownCtrl into its own module
