+++ lib/dropdown/index.js
@@ -2,32 +2,9 @@ import angular  from 'angular';
 import box      from '../box/index';
 import icons    from '../icons/index';
 
-var dropdown = angular.module('guDropdown', ['guBox', 'guIcons']);
+import DropdownCtrl from './DropdownCtrl';
 
-dropdown.controller('guDropdownCtrl', [
-  '$scope', function dropdownController(scope){
-
-      function getClassName() {
-        return scope.state === 'open' ? 'dropdown open' : 'dropdown closed';
-      }
-
-      function assignClass() {
-        scope.className = getClassName();
-      }
-
-      function onTriggerClicked() {
-        //toggle state
-        scope.state = (scope.state === 'open') ? 'closed' : 'open';
-        assignClass();
-      }
-
-      //when the dropdown trigger is clicked
-      this.onTriggerClicked = onTriggerClicked;
-
-      //assign the starting class
-      assignClass();
-  }
-])
+var dropdown = angular.module('guDropdown', ['guBox', 'guIcons', 'guDropdownCtrl']);
 
 dropdown.directive('guDropdown', function dropdownDirective(){
   return {
